Port to Teensy
It looks like a single Teensy can drive all the LEDs itself, as well as
interface to the accelerometers. Doing everything on the one CPU will make
life much simpler. Port the game code to run in an Arduino environment
(which can run on the Teensy) and add LED and accelerometer support.

This compiles, but is entirely untested since I have no access to the
latest Teensy HW, nor the LED array, nor the accelerometer boards! As
such, it's probably riddled with stupid bugs.

Signed-off-by: Stephen Warren <swarren@wwwdotorg.org>